fix other tests
