+++ test/jb_payment_terminal/distribute_payouts_of.test.js
@@ -1,19 +1,19 @@
 import { expect } from 'chai';
 import { ethers } from 'hardhat';
 import { deployMockContract } from '@ethereum-waffle/mock-contract';
-import { makeSplits, packFundingCycleMetadata, setBalance } from '../../helpers/utils.js';
-import errors from '../../helpers/errors.json';
+import { makeSplits, packFundingCycleMetadata, setBalance } from '../helpers/utils.js';
+import errors from '../helpers/errors.json';
 
-import jbAllocator from '../../../artifacts/contracts/interfaces/IJBSplitAllocator.sol/IJBSplitAllocator.json';
-import jbDirectory from '../../../artifacts/contracts/JBDirectory.sol/JBDirectory.json';
-import JBETHPaymentTerminal from '../../../artifacts/contracts/JBETHPaymentTerminal.sol/JBETHPaymentTerminal.json';
-import jbPaymentTerminalStore from '../../../artifacts/contracts/JBSingleTokenPaymentTerminalStore.sol/JBSingleTokenPaymentTerminalStore.json';
-import jbFeeGauge from '../../../artifacts/contracts/interfaces/IJBFeeGauge.sol/IJBFeeGauge.json';
-import jbOperatoreStore from '../../../artifacts/contracts/JBOperatorStore.sol/JBOperatorStore.json';
-import jbProjects from '../../../artifacts/contracts/JBProjects.sol/JBProjects.json';
-import jbSplitsStore from '../../../artifacts/contracts/JBSplitsStore.sol/JBSplitsStore.json';
-import jbPrices from '../../../artifacts/contracts/JBPrices.sol/JBPrices.json';
-import IERC20Metadata from '../../../artifacts/@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol/IERC20Metadata.json';
+import jbAllocator from '../../artifacts/contracts/interfaces/IJBSplitAllocator.sol/IJBSplitAllocator.json';
+import jbDirectory from '../../artifacts/contracts/JBDirectory.sol/JBDirectory.json';
+import JBETHPaymentTerminal from '../../artifacts/contracts/JBETHPaymentTerminal.sol/JBETHPaymentTerminal.json';
+import jbPaymentTerminalStore from '../../artifacts/contracts/JBSingleTokenPaymentTerminalStore.sol/JBSingleTokenPaymentTerminalStore.json';
+import jbFeeGauge from '../../artifacts/contracts/interfaces/IJBFeeGauge.sol/IJBFeeGauge.json';
+import jbOperatoreStore from '../../artifacts/contracts/JBOperatorStore.sol/JBOperatorStore.json';
+import jbProjects from '../../artifacts/contracts/JBProjects.sol/JBProjects.json';
+import jbSplitsStore from '../../artifacts/contracts/JBSplitsStore.sol/JBSplitsStore.json';
+import jbPrices from '../../artifacts/contracts/JBPrices.sol/JBPrices.json';
+import IERC20Metadata from '../../artifacts/@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol/IERC20Metadata.json';
 
 describe('JBPayoutRedemptionPaymentTerminal::distributePayoutsOf(...)', function () {
   const PLATFORM_PROJECT_ID = 1;
@@ -1977,6 +1977,10 @@ describe('JBPayoutRedemptionPaymentTerminal::distributePayoutsOf(...)', function
 
     await Promise.all(
       splits.map(async (split) => {
+        await fakeToken.mock.allowance
+          .withArgs(jbErc20PaymentTerminal.address, mockJbAllocator.address)
+          .returns(0);
+
         await fakeToken.mock.approve
           .withArgs(
             mockJbAllocator.address,
@@ -2389,6 +2393,10 @@ describe('JBPayoutRedemptionPaymentTerminal::distributePayoutsOf(...)', function
 
     await Promise.all(
       splits.map(async (split) => {
+        await fakeToken.mock.allowance
+          .withArgs(jbErc20PaymentTerminal.address, mockJbEthPaymentTerminal.address)
+          .returns(0);
+
         await fakeToken.mock.approve
           .withArgs(
             mockJbEthPaymentTerminal.address,
@@ -2764,6 +2772,134 @@ describe('JBPayoutRedemptionPaymentTerminal::distributePayoutsOf(...)', function
         /*_feeAmount*/ 0,
         /*_leftoverDistributionAmount*/ AMOUNT_DISTRIBUTED -
           ((AMOUNT_DISTRIBUTED * PERCENT) / SPLITS_TOTAL_PERCENT) * splits.length,
+        MEMO,
+        caller.address,
+      );
+  });
+
+  it('Should not take a fee of 1 if leftover is 1 (fix rounding error)', async function () {
+    const {
+      projectOwner,
+      terminalOwner,
+      caller,
+      beneficiaryOne,
+      beneficiaryTwo,
+      jbEthPaymentTerminal,
+      timestamp,
+      mockJbSplitsStore,
+      mockJbDirectory,
+      mockJBPaymentTerminalStore,
+    } = await setup();
+
+    const AMOUNT_TO_DISTRIBUTE = 101;
+
+    const splits = makeSplits({
+      count: 2,
+      projectId: OTHER_PROJECT_ID,
+      beneficiary: [beneficiaryOne.address, beneficiaryTwo.address],
+    });
+
+    await mockJbSplitsStore.mock.splitsOf
+      .withArgs(PROJECT_ID, timestamp, ETH_PAYOUT_INDEX)
+      .returns(splits);
+
+    await mockJbDirectory.mock.primaryTerminalOf
+      .withArgs(OTHER_PROJECT_ID, ETH_ADDRESS)
+      .returns(jbEthPaymentTerminal.address);
+
+    // ETH distribution
+    await mockJBPaymentTerminalStore.mock.recordDistributionFor
+      .withArgs(PROJECT_ID, AMOUNT_TO_DISTRIBUTE, CURRENCY)
+      .returns(fundingCycle, AMOUNT_TO_DISTRIBUTE);
+
+    await Promise.all(
+      splits.map(async (split) => {
+        await mockJBPaymentTerminalStore.mock.recordPaymentFrom
+          // .withArgs(
+          //   jbEthPaymentTerminal.address,
+          //   [
+          //     /*token*/ '0x000000000000000000000000000000000000eeee',
+          //     /*amount paid*/ Math.floor(
+          //       (AMOUNT_TO_DISTRIBUTE * split.percent) / SPLITS_TOTAL_PERCENT,
+          //     ),
+          //     /*decimal*/ 18,
+          //     CURRENCY,
+          //   ],
+          //   split.projectId,
+          //   CURRENCY,
+          //   split.beneficiary,
+          //   '',
+          //   ethers.utils.hexZeroPad(ethers.utils.hexlify(PROJECT_ID), 32),
+          // )
+          .returns(fundingCycle, /*count*/ 0, /* delegateAllocation */ [], '');
+      }),
+    );
+
+    // POC Bug: pay 1 wei of fee =
+    await mockJBPaymentTerminalStore.mock.recordPaymentFrom
+      .withArgs(
+        jbEthPaymentTerminal.address,
+        [
+          /*token*/ '0x000000000000000000000000000000000000eeee',
+          /*amount paid*/ 1,
+          /*decimal*/ 18,
+          CURRENCY,
+        ],
+        1,
+        CURRENCY,
+        projectOwner.address,
+        '',
+        ethers.utils.hexZeroPad(ethers.utils.hexlify(PROJECT_ID), 32),
+      )
+      .returns(fundingCycle, /*count*/ 0, /* delegateAllocation */ [], '');
+
+    let tx = await jbEthPaymentTerminal
+      .connect(caller)
+      .distributePayoutsOf(
+        PROJECT_ID,
+        AMOUNT_TO_DISTRIBUTE,
+        ETH_PAYOUT_INDEX,
+        ethers.constants.AddressZero,
+        AMOUNT_TO_DISTRIBUTE,
+        MEMO,
+      );
+
+    await Promise.all(
+      splits.map(async (split) => {
+        await expect(tx)
+          .to.emit(jbEthPaymentTerminal, 'DistributeToPayoutSplit')
+          .withArgs(
+            PROJECT_ID,
+            /*_fundingCycle.configuration*/ timestamp,
+            ETH_PAYOUT_INDEX,
+            [
+              split.preferClaimed,
+              split.preferAddToBalance,
+              split.percent,
+              split.projectId,
+              split.beneficiary,
+              split.lockedUntil,
+              split.allocator,
+            ],
+            /*payoutAmount*/ Math.floor(
+              (AMOUNT_TO_DISTRIBUTE * split.percent) / SPLITS_TOTAL_PERCENT,
+            ),
+            caller.address,
+          );
+      }),
+    );
+
+    await expect(tx)
+      .to.emit(jbEthPaymentTerminal, 'DistributePayouts')
+      .withArgs(
+        /*_fundingCycle.configuration*/ timestamp,
+        /*_fundingCycle.number*/ 1,
+        PROJECT_ID,
+        projectOwner.address,
+        /*_amount*/ AMOUNT_TO_DISTRIBUTE,
+        /*_distributedAmount*/ AMOUNT_TO_DISTRIBUTE,
+        /*_feeAmount*/ 0,
+        /*_leftoverDistributionAmount*/ 1,
         MEMO,
         caller.address,
       );
